Don't reset target if out-of-sight enemy is present at target position
If you aimed outside of the field of view and an enemy was present on
that space, the next target default position would be reset to the
player, instead of staying in place like it should.  This was not only
annoying, but also leaked information about enemy positions.

(cherry picked from commit e487c603c73a1a89c47d6c9fdf7e69214cbf582f)